backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-google-genai
chromadb
pandas
pydantic
langchain-community
